Implement smoked, a service to read hardware events
SmokeD, intended to run as a privileged daemon, observers the state of
the hardware and publishes relevant changes to a system wide message
bus. The intention is for less privileged processes to subscribe to this
bus to read hardware state. Eventually a mechanism will be added to
allow unprivileged processes to not only read from hardware, but write
to it as well.