Enable SmokeD to run w/o caring about threads
Each Observable now runs its own event loop in its own thread. A default
event loop that blocks in a CPU efficient way is provided for Observables that don't require such functionality. The event loop starts as soon as an observer is added and stops when all observers are removed.